Fixed migration order refactor side effect
* Moved Configuration and S3LIMS into limsrow module
  to maintain migration order
* Fixed couples of lints
* QA and improved refactor work from #322 #341
