--- data_portal/models/configuration.py
@@ -1,55 +1,0 @@
-import logging
-
-from typing import Union
-
-from django.db import models
-
-logger = logging.getLogger(__name__)
-
-class Configuration(models.Model):
-    """
-    Model that stores a configuration value
-    Currently not used; but could be used for cases when we want to record the state of some data - e.g. LIMS file.
-    """
-    id = models.BigAutoField(primary_key=True)
-    name = models.CharField(max_length=255, unique=True)
-    value = models.TextField()
-
-    @staticmethod
-    def get(name: str) -> Union['Configuration', None]:
-        """
-        Get configuration object by name. None if not exists
-        """
-        query_set = Configuration.objects.filter(name=name)
-        if query_set.exists():
-            return query_set.get()
-        return None
-
-    @staticmethod
-    def set(name: str, val: str) -> None:
-        """
-        Set config value by name. Will create one if not exist.
-        """
-        config = Configuration.get(name)
-
-        if config:
-            config.value = val
-        else:
-            config = Configuration(name=name, value=val)
-
-        config.save()
-
-    @staticmethod
-    def same_or_update(name: str, val: str):
-        """
-        Compares val with current config value if exist.
-        Update if config not exist or the two values are not equal.
-        :return True if two values are same; False if they are different or current val does not exist.
-        """
-        config = Configuration.get(name)
-
-        if config is None or config.value != val:
-            Configuration.set(name, val)
-            return False
-
-        return True
