+++ data_portal/models/libraryrun.py
@@ -7,6 +7,7 @@ from data_portal.models.workflow import Workflow
 
 logger = logging.getLogger(__name__)
 
+
 class LibraryRunManager(models.Manager):
 
     def get_by_keyword(self, **kwargs) -> QuerySet:
