+++ data_portal/models/tests/test_sequence.py
@@ -1,14 +1,16 @@
 import logging
 
-from django.test import TestCase
-from data_portal.models.sequence import Sequence
 from django.core.exceptions import ObjectDoesNotExist
+from django.test import TestCase
+
+from data_portal.models.sequence import Sequence
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
 
 class SequenceTestCase(TestCase):
+
     def setUp(self):
         logger.info('Create Object data')
         Sequence.objects.create(
@@ -40,24 +42,3 @@ class SequenceTestCase(TestCase):
             Sequence.objects.get(status='Complete')
         except ObjectDoesNotExist:
             logger.info(f"Raised ObjectDoesNotExist")
-
-    def test_get_api(self):
-
-        # Get sequencee list
-        logger.info('Get sequence API')
-        response = self.client.get('/sequence/')
-        self.assertEqual(response.status_code, 200, 'Ok status response is expected')
-
-        logger.info('Check if API return result')
-        result_response = response.data['results']
-        self.assertGreater(len(result_response), 0, 'A result is expected')
-
-        logger.info('Check if unique data has a signle entry')
-        response = self.client.get('/sequence/?instrument_run_id=191213_A00000_00000_A000000000&run_id=r.AAAAAA')
-        results_response = response.data['results']
-        self.assertEqual(len(results_response), 1, 'Single result is expected for unique data')
-
-        logger.info('Check if wrong parameter')
-        response = self.client.get('/sequence/?lib_id=LBR0001')
-        results_response = response.data['results']
-        self.assertEqual(len(results_response), 0, 'No result is expected for wrong parameter')
