+++ data_portal/viewsets/libraryrun.py
@@ -19,6 +19,7 @@ logger = logging.getLogger()
 LIBRARY_RUN_SEARCH_FIELDS = ["id", "library_id", "instrument_run_id", "run_id", "lane", "override_cycles",
                              "coverage_yield", "qc_pass", "qc_status", "valid_for_analysis"]
 
+
 class LibraryRunViewSet(ReadOnlyModelViewSet):
     serializer_class = LibraryRunModelSerializer
     pagination_class = StandardResultsSetPagination
