+++ data_portal/viewsets/tests/__init__.py
@@ -1,0 +1,4 @@
+import logging
+
+_logger = logging.getLogger()
+_logger.setLevel(logging.INFO)
