+++ data_portal/viewsets/utils.py
@@ -1,6 +1,8 @@
 from rest_framework.response import Response
 
 from utils import libs3
+
+
 def _error_response(message, status_code=400, err=None) -> Response:
     data = {'error': message}
     if err:
