+++ data_processors/s3/services.py
@@ -16,9 +16,8 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.db import transaction
 from django.db.models import ExpressionWrapper, Value, CharField, Q, F
 
+from data_portal.models.limsrow import LIMSRow, S3LIMS
 from data_portal.models.s3object import S3Object
-from data_portal.models.limsrow import LIMSRow
-from data_portal.models.s3lims import S3LIMS
 from data_processors.const import S3EventRecord
 from utils import libs3
